Move Attributes & Behaviors to Data
